Core/Compilation: Added rvalue references.
The reason why we don't pass it by reference is that we can't pass a
non-const rvalue by reference (please learn yourself
http://stackoverflow.com/questions/4084053/passing-non-const-references-to-rvalues-in-c)
and most of the time the method NextRow is used and non-const. So the
only correct way to pass it in a fast way is to use rvalue references
(http://thbecker.net/articles/rvalue_references/section_01.html).