Core/Spells: Refresh timers of periodic heal/damage has change to Wod :
"Blizzard will be making the same type of refresh mechanic that is available for Warlocks available for all HOT and DOT effects, just on a smaller scale.  The new mechanic is that you can refresh them anywhere in the last 30% of the normal duration and keep the initial full effect.  This means your new application could be up to 130% duration of normal."